Revert "console: only allow console messages after initialization"
This reverts commit 60320182d011a8c31793f833f36f65cd5e8b50df.

This commit for some reason leads to the following message on A8V-E SE:
Romstage handoff structure not added!
(which makes the boot hang)

Likely this reverted commit is not the cause for this, but there is some
corruption as indicated by
BSP overran lower stack boundary.  Undefined behaviour may result!
and we were just lucky, but this commit somehow make the breakage
visible.

Change-Id: Iaf15d871e3a4a0bff6ff38aa717dfdaf1b9e96ef